ref(console): list the commands once, not once per registry
The same nineteen commands were written out three times:
`ConsoleProvider::commands()`, which `vendor/bin/gacela` reads, and a
`GacelaCommands` in each bridge. Adding a command meant remembering all
three, and three commands had already been missing from the bridges once.

Guards were written after that -- `RegisteredCommandsTest` for the CLI and
a `GacelaCommandsTest` per bridge, all reading the command directory rather
than restating the list. They work, and they stay. They turn a forgotten
registry into a failing build; they do not make the list one list.

`CommandCatalog` is that list. It also absorbs the one special case the
three registries each spelled out: `InitCommand` is the only command taking
a constructor argument, and the two bridges passed `''` purely to read a
name off it.

The guards now cover a different mistake -- a command class in the
directory and in no list at all -- and were checked to still fire for it:
adding an unlisted command fails all three, each naming its own surface.
